Update sync.js
Updated the comment. As it was written, it felt as if this library performed the actual sending of the message contained in `encoder`. However, this library is network agnostic. 

Rather, it seems, it's up to the server using this lib whether it wants to send the message or not, and should not send it if it detects the message is empty (as seen [here](https://github.com/yjs/y-websocket/blob/625a9d52ba8bcff90a941d51ae151869e220f368/bin/utils.js#L171)).